refactor: separate Character class, add hono, set up hono endpoints, connect frontend loader
