+++ apps/backend/package.json
@@ -19,6 +19,8 @@
 		"wrangler": "^3.72.0"
 	},
 	"dependencies": {
+		"@hono/zod-validator": "^0.2.2",
+		"hono": "^4.5.6",
 		"zod": "^3.23.8"
 	}
 }
