+++ apps/backend/worker-configuration.d.ts
@@ -2,5 +2,5 @@
 // After adding bindings to `wrangler.toml`, regenerate this interface via `npm run cf-typegen`
 interface Env {
 	MY_DURABLE_OBJECT: DurableObjectNamespace;
+	AI: Ai;
 }
-
