--- apps/frontend/hello-world.ts
@@ -1,4 +1,0 @@
-export const onRequest: PagesFunction<Env> = (context) => {
-  console.log(context.env.ASSETS);
-  return new Response("Hello, world!");
-};
